Fix component names
