+++ client/src/Components/AppList/index.jsx
@@ -1,15 +1,15 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 
-import AppListing from '../AppListing/AppListing.jsx';
-import AppModal from '../AppModal/AppModal.jsx';
+import AppListing from '../AppListing';
+import AppModal from '../AppModal';
 
-import './AppList.scss';
+import './applist.scss';
 
 export default function AppList() {
   const [position, setPosition] = useState('');
   const [company, setCompany] = useState('');
-  
+
   const [appModal, toggleAppModal] = useState(false);
   const [apps, setApps] = useState([]);
 
@@ -21,14 +21,12 @@ export default function AppList() {
     <div className="appList-container">
       <h1 className="appList-title">Applications</h1>
       <button onClick={() => toggleAppModal(true)} className="appList-btn-add">
-      + Add App
+        + Add App
       </button>
       {apps.map(app => {
         const { position, company, id } = app;
         console.log(app);
-        return (
-          <AppListing key={id} position={position} company={company} />
-        );
+        return <AppListing key={id} position={position} company={company} />;
       })}
       <AppModal setApps={setApps} appModal={appModal} toggleAppModal={toggleAppModal} />
     </div>
